Moved more files around to reduce code line size
