+++ c_version/src/main.c
@@ -5,8 +5,6 @@
 #define ALERT_TIMES 1
 #define DISPLAY_OUTPUT 0
 
-void usage();
-
 int main(int argc, char *argv[]){
     int quietMode = 0;
     int dryRunMode = 0;
@@ -44,32 +42,3 @@ int main(int argc, char *argv[]){
         runProgram(argSwitch, DISPLAY_OUTPUT);
 }
 
-void usage(void){
-    printf("usage:\n"
-           "Countdown to the second, minute or hour,  or by a clock time\n"
-           "                                                            \n"
-           "timer [seconds]                                             \n"
-           "timer [minutes] [seconds]                                   \n"
-           "timer [hours] [minutes] [seconds]                           \n"
-           "timer [clockAM/PM]                                          \n"
-           "timer -p                                                    \n"
-           "                                                            \n"
-           "Switches                                                    \n"
-           "-p              prompt time input                           \n"
-           "--quiet         Doesn't play the alert                      \n"
-           "--dry-run         Doesn't play the alert                    \n"
-           "                                                            \n"
-           "EXAMPLE: 2 hours, 10 minutes, 1 second                      \n"
-           "timer 2 10 1                                                \n"
-           "                                                            \n"
-           "EXAMPLE: Timer for 5:30 24 hours                            \n"
-           "timer 5:30                                                  \n"
-           "                                                            \n"
-           "EXAMPLE: Timer for 5:30am 12 hour clock                     \n"
-           "timer 5:30am                                                \n"
-           "                                                            \n"
-           "EXAMPLE: Timer for 5pm 12 hour clock                        \n"
-           "timer 5pm                                                   \n"
-           "                                                            \n"
-           );
-}
